### 2024-09-20 #7247 @editorialbot
[PRE REVIEW]: madupite: A High-Performance Distributed Solver for Large-Scale Markov Decision Processes
**Submitting author:** @gmatilde (<a href="http://orcid.org/0000-0001-8615-6214">Matilde Gargiani</a>)
**Repository:** https://github.com/madupite/madupite
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
