### 2024-09-25 165c9d3
adding missing DOIs
