+++ joss/paper.bib
@@ -7,6 +7,7 @@
     number       = 2,
     pages        = {11249--11254},
     issn         = {2405-8963},
+    doi          = {https://doi.org/10.1016/j.ifacol.2023.10.316},
     url          = {https://www.sciencedirect.com/science/article/pii/S2405896323006833},
     annote       = {22nd IFAC World Congress}
 }
@@ -72,6 +73,7 @@
     year         = 2017,
     publisher    = {Springer},
     pages        = {189--222},
+    doi          = {10.1007/978-3-319-47766-4_6},
     url          = {https://EconPapers.repec.org/RePEc:spr:isochp:978-3-319-47766-4_6},
     chapter      = 6
 }
@@ -115,6 +117,7 @@
     booktitle    = {Modern Software Tools in Scientific Computing},
     publisher    = {Birkh{\"{a}}user Press},
     pages        = {163--202},
+    doi          = {10.1007/978-1-4612-1986-6_8},
     editor       = {E. Arge and A.~M. Bruaset and H.~P. Langtangen}
 }
 @article{mills2021,
