### 2024-09-25 e504127
correcting some DOIs
