+++ joss/paper.bib
@@ -7,7 +7,7 @@
     number       = 2,
     pages        = {11249--11254},
     issn         = {2405-8963},
-    doi          = {https://doi.org/10.1016/j.ifacol.2023.10.316},
+    doi          = {10.1016/j.ifacol.2023.10.316},
     url          = {https://www.sciencedirect.com/science/article/pii/S2405896323006833},
     annote       = {22nd IFAC World Congress}
 }
@@ -25,7 +25,7 @@
     volume       = 69,
     number       = 1,
     pages        = {455--462},
-    url          = {https://doi.org/10.1109/TAC.2023.3270060}
+    url          = {10.1109/TAC.2023.3270060}
 }
 @software{mdp_solver,
     title        = {areenberg/MDPSolver: MDPSolver v0.9.7},
