+++ joss/paper.bib
@@ -25,7 +25,8 @@
     volume       = 69,
     number       = 1,
     pages        = {455--462},
-    url          = {10.1109/TAC.2023.3270060}
+    url          = {https://doi.org/10.1109/TAC.2023.3270060},
+    doi          = {10.1109/TAC.2023.3270060}
 }
 @software{mdp_solver,
     title        = {areenberg/MDPSolver: MDPSolver v0.9.7},
