### 2024-09-25 #7247 @logological
> 👋 @logological - Do you think that you could edit this JOSS submission?

It's a bit outside my area but I could take it on if there's no one more suitable available.  I'm at a conference this week but could start the editing process next week.
