### 2024-10-24 #7247 @logological
@alberto743 @lukeolson @bhttchr6 We've received a submission entitled "madupite: A High-Performance Distributed Solver for Large-Scale Markov Decision Processes". Would you be interested in reviewing it for the Journal of Open Source Software?
