### 2024-10-24 #7247 @alberto743
@logological Thank you for considering me as a potential reviewer for this work. I already have two other pending reviews for JOSS. Realistically, I will not be available for additional reviews before the beginning of February 2025. If you do not find other reviewers, come back to me in about two months.
