### 2024-10-24 #7247 @bhttchr6
Hello,
Yes, I am interested in reviewing the article.
Thanks,
Regards,
Anurag

On Thu, Oct 24, 2024 at 12:31 PM Tristan Miller ***@***.***>
wrote:

> @alberto743 <https://github.com/alberto743> @lukeolson
> <https://github.com/lukeolson> @bhttchr6 <https://github.com/bhttchr6>
> We've received a submission entitled "madupite: A High-Performance
> Distributed Solver for Large-Scale Markov Decision Processes". Would you be
> interested in reviewing it for the Journal of Open Source Software?
>
