### 2024-10-24 #7247 @lukeolson
@logological - yes, happy to review.
