### 2024-11-04 #7411 @gmatilde
Hi @victorapm, thanks for your message! I was also the project leader, ideator and I did design the solution methods which the solver uses.
