### 2024-11-21 #7411 @lukeolson
@gmatilde -- A few suggestions for the paper.  Please consider these minor and only suggestions to (hopefully!) help improve the flow:
- line 30: `solving gigantic size MDPs` -> `solving MPDs that are large in size`
- line 36: `plain Python and this results in poor scalability` -> `plain Python, which is limited in scalability due to the lack of parallel support.`
- line 39: `On the other hand, the solution methods available are` -> `Even so, mdpsolver is not fully featured.  The solution methods are`
- line 41: `On the other hand, ' -> `In addition, `
- line 44: remove the sentence `Finally, ...`
- line 61: what is meant by `general enough to embrace`?
- line 78: remove `despite being developed specifically for solving partial differential equations`
- line 82: remove `which is one of the innovative key-features of our solver with respect to its competitors`
