### 2024-11-21 #7411 @lukeolson
@gmatilde .  I have completed a review of each of the items in the checklist.  A few questions popped up as I went through the items.  Here they are, each with a recommendation:
- **installation** madupite is limited to linux (https://github.com/madupite/madupite/blob/dbc3aff095c62f84ff8e48dff89da1df3fccea1c/environment.yml#L6).  **recommendation** This should be noted in the instructions.  If installation instructions for a mac could be added, even better.
- **testing** madupite does use pytest and CI.  However, from the tests it appears that only the execution is exercised.  It is unclear that algorithm correctness is verified. (https://github.com/madupite/madupite/tree/main/tests).  **recommendation** Implement at least one model problem (say from the examples) in testing.
- **performance** There are claims of scalability with the software.  **recommendation** consider adding an example to the paper on scalability.
- **scholarly effort** From the JOSS guidelines, the commit history (~1.5 years), number of commits (~526), total lines of code (>2000 in .cpp and .h files).  So I think it meets the standard for substantial scholarly effort.  **recommendation** In the paper or in the documentation it may be helpful to sketch out the overall design -- I believe much of this is in `mdp_algorithm.cpp`.
