### 2024-11-21 #7411 @lukeolson
@logological done here, with some suggestions above.
