### 2025-02-04 #7411 @bhttchr6
@gmatilde Few minor suggestions for the paper:
1. Line 19-21: add a reference for the statement.
2. Line 30: "solving gigantic size", consider replacing it with a more quantitative number/range.
3. Line 30-31: Consider " To exploit modern computational resources, high performance software packages which can efficiently distribute computation as well as scale adequately are needed" .
4. Line 47: consider revising the statement to make the vision/goal clearer to the reader.
5. Line 68:" the great flexibility of madupite **is that** on the algorithmic side **it** relies...."
6. The performance of the software is not clearly demonstrated, please consider adding examples to demonstrate the same.
