### 2025-02-04 #7411 @victorapm
I'm also done with my review. This is a nice work! I just wanted to reiterate on the performance comments that the other reviewers brought up. It would be great to have an example in the paper covering this topic. Thank you!
