### 2025-02-04 #7411 @rjs02
Hi all, one of the other authors writing since @gmatilde  is out of office until next week. I’d like to thank you all @lukeolson, @victorapm and @bhttchr6 for your efforts in reviewing madupite and the kind feedback. We have been working on a follow-up paper focusing solely on performance benchmarks and scalability which is ready for submission soon. Regarding the suggestions on the JOSS paper, we’ll try to have everything done by end of next week so we can finish this reviewing process. :)
