### 2025-02-20 #7411 @gmatilde
@lukeolson @bhttchr6 @victorapm Thank you for your time and valuable feedback on our work! We’re delighted to hear that madupite was well received. We have worked to incorporate your comments and believe we have addressed the main concerns. With these improvements, we hope it is now an even more useful and user-friendly tool for the research community.

Here are some additional clarifications:

Regarding **performance** and **benchmarks**, we chose not to include them due to space limitations. In line with JOSS guidelines, we kept the paper concise and general, and we had already reached the maximum allowed length. Properly introducing, contextualizing, and discussing benchmarks would have required more space than available. However, we recognize the importance of this aspect and consider it a valuable contribution. Therefore, we are currently working on a dedicated paper focused on this topic, with a draft expected to be ready within the next two months. This new paper does not overlap with the current work but rather complements it, making it an appropriate and independent contribution. That said, users can explore various ready-to-use examples in the `examples` folder to study madupite's scalability and performance. Some case studies, such as the inverted pendulum, are also detailed in the official madupite documentation.

Regarding **testing**, the current pipeline includes small MDPs to verify the correctness of the installation and overall code functionality. These tests can be found in the `examples/ci_test` folder.

Lastly, we have now added a `CONTRIBUTING.md` file that provides a detailed description of the repository's layout and organization for improved clarity.

We sincerely appreciate your thoughtful feedback and suggestions. Your insights have helped us refine and improve madupite, making it an even more robust and user-friendly tool. We hope that our clarifications address your concerns and that the provided resources will be useful. Thank you again for your time and valuable contributions to our work!
