### 2025-02-24 #7411 @logological
Notes on the DOI checks, with a few requests to the authors for cleanup:

> - No DOI given, and none found for title: Inexact Policy Iteration Methods for Large-Scale M...

@gmatilde: Please add the DOI for this paper: 10.48550/arXiv.2404.06136

> - No DOI given, and none found for title: Markov decision processes with applications to fin...

OK; the paper cites the softcover edition.  (Only the electronic edition has a DOI.)

> - No DOI given, and none found for title: Handbook of Markov Decision Processes: Methods and...

OK; the paper cites Kluwer's print edition.  (Only the Springer electronic edition has a DOI.)

> - No DOI given, and none found for title: PETSc Web page

OK; the website has no DOI.

> - No DOI given, and none found for title: nanobind: tiny and efficient C++/Python bindings

@gmatilde: This reference is missing the URL.  Could you please add it?

> - No DOI given, and none found for title: Dynamic Programming and Optimal Control, Vol. II

OK; confirmed 3rd edition cited has no DOI.

> - No DOI given, and none found for title: Reinforcement Learning: An Introduction

@gmatilde: Consider adding the ISBN for this book: 978-0-262-03924-6

> ❌ MISSING DOIs
>
> - 10.1201/9780429123641-3 may be a valid DOI for title: Dynamic Programming

OK; the bot is wrong here as this is not the DOI for the book.
