### 2025-02-26 #7411 @gmatilde
@logological thanks for your feedback! the version is v1.0.0 and it's archived here: https://doi.org/10.5281/zenodo.14930981
