### 2025-02-28 #7411 @logological
@gmatilde: Thanks for fixing the references.  There's no need to archive on Zenodo yet as we're still conducting some final checks to the manuscript.  Could you please now address the following issues with the writing?

On page 2, the sentence beginning "madupite is high-performance distributed solver..." is exceedingly long, as well as awkwardly and ambiguously structured.  It's made up of a list of three features of madupite, the first and third of which are phrased as fully independent clauses with a subject and a verb, but the second of which clips the subject.  Furthermore, it's not clear whether the second and third items in the list attach to the main subject of the sentence ("madupite") or to the anaphor of "which" ("a high-performance distributed solver").  Assuming that it reflects your intended meaning, could I suggest that you rephrase and split up this sentence as follows?

> madupite is a high-performance distributed solver that is capable of efficiently distributing the memory load and computation, and that comes with a wide range of choices for solution methods enabling the user to select the one that is best tailored to its specific application.  Furthermore, madupite's core is in C++, but it is equipped with a user-friendly Python API to enable its deployment to a broader range of users.

Also on page 2, change "The interested readers" to "Interested readers".  (Use of the definite article is not indicated here since the paper never previously referred to a specific group of interested readers.)

In the first paragraph of page 3, consider moving the citation to the Balay et al. resources to the next sentence, immediately after the word "PETSc".  In its current position, it looks like the citation is supporting the main claim of the sentence ("The core of madupite is written in C++ and relies on PETSc") rather than simply referring the reader to further information on PETSc.

On page 3, change "code-examples" to "code examples".
