### 2025-03-28 #7411 @logological
@gmatilde Thanks for applying the copy-edits.  Since you've included the paper source in your Zenodo archive, that version is now out of sync with the to-be-published one in the GitHub repository.  Could you produce a new version of the archive on Zenodo with the latest version of the paper (or without the paper at all)?  Also, could you please ensure that the Zenodo archive title exactly matches the title of the paper?
