### 2025-03-29 #7411 @rjs02
Hi @logological, I updated the Zenodo archive (removed the paper and changed the title). The new DOI is https://doi.org/10.5281/zenodo.15105482 and v1.0.1
