### 2025-03-31 #7411 @logological
@rjs02 @gmatilde Thanks.  I note that the licence listed in the Zenodo archive (Creative Commons Attribution 4.0 International) differs from the one listed in the GitHub repository (MIT).  I assume it's the Zenodo archive licence that's in error, since CC licences are generally not appropriate for software.  Could you please change the Zenodo archive licence to match the one in the GitHub repository?
