### 2025-04-01 #7411 @danielskatz
@gmatilde - I've suggested a few small changes in https://github.com/madupite/madupite/pull/8 - please merge this, or let me know what you disagree with.

In addition, in the first sentence, you write "We propose madupite", but this is not a proposal, this is software you have built. Perhaps you might change this to "In this paper, we describe madupite"?
