### 2025-04-02 #7411 @gmatilde
@danielskatz Thanks a lot for your time! I have merged your changes to madupite and I included the minor correction in the paper. Please let us know in case you would like to modify something else in the code or the main body of the paper.
