### 2025-04-02 #7411 @danielskatz
Congratulations to @gmatilde (Matilde Gargiani) and co-authors on your publication!!

JOSS depends on volunteers and couldn't do this without you
