### 2025-04-02 #7411 @bhttchr6
@gmatilde Congratulations!
@danielskatz @logological Looking forward to assisting with more reviews for JOSS
